+++ core/src/main/java/org/elasticsearch/common/geo/builders/MultiPointBuilder.java
@@ -31,6 +31,7 @@ import java.util.List;
 
 public class MultiPointBuilder extends PointCollection<MultiPointBuilder> {
 
+
     public static final GeoShapeType TYPE = GeoShapeType.MULTIPOINT;
 
     @Override
