+++ core/src/main/java/org/elasticsearch/common/geo/builders/MultiPolygonBuilder.java
@@ -33,7 +33,7 @@ public class MultiPolygonBuilder extends ShapeBuilder {
 
     public static final GeoShapeType TYPE = GeoShapeType.MULTIPOLYGON;
 
-    protected final ArrayList<BasePolygonBuilder<?>> polygons = new ArrayList<>();
+    protected final ArrayList<PolygonBuilder> polygons = new ArrayList<>();
 
     public MultiPolygonBuilder() {
         this(Orientation.RIGHT);
@@ -43,7 +43,7 @@ public class MultiPolygonBuilder extends ShapeBuilder {
         super(orientation);
     }
 
-    public MultiPolygonBuilder polygon(BasePolygonBuilder<?> polygon) {
+    public MultiPolygonBuilder polygon(PolygonBuilder polygon) {
         this.polygons.add(polygon);
         return this;
     }
@@ -53,7 +53,7 @@ public class MultiPolygonBuilder extends ShapeBuilder {
         builder.startObject();
         builder.field(FIELD_TYPE, TYPE.shapename);
         builder.startArray(FIELD_COORDINATES);
-        for(BasePolygonBuilder<?> polygon : polygons) {
+        for(PolygonBuilder polygon : polygons) {
             builder.startArray();
             polygon.coordinatesArray(builder, params);
             builder.endArray();
@@ -73,13 +73,13 @@ public class MultiPolygonBuilder extends ShapeBuilder {
         List<Shape> shapes = new ArrayList<>(this.polygons.size());
 
         if(wrapdateline) {
-            for (BasePolygonBuilder<?> polygon : this.polygons) {
+            for (PolygonBuilder polygon : this.polygons) {
                 for(Coordinate[][] part : polygon.coordinates()) {
                     shapes.add(jtsGeometry(PolygonBuilder.polygon(FACTORY, part)));
                 }
             }
         } else {
-            for (BasePolygonBuilder<?> polygon : this.polygons) {
+            for (PolygonBuilder polygon : this.polygons) {
                 shapes.add(jtsGeometry(polygon.toPolygon(FACTORY)));
             }
         }
