+++ core/src/main/java/org/elasticsearch/common/geo/builders/PointCollection.java
@@ -34,7 +34,6 @@ import com.vividsolutions.jts.geom.Coordinate;
 public abstract class PointCollection<E extends PointCollection<E>> extends ShapeBuilder {
 
     protected final ArrayList<Coordinate> points;
-    protected boolean translated = false;
 
     protected PointCollection() {
         this(new ArrayList<Coordinate>());
